refactor: rename utils to utilities
