+++ subgraphs/venus-governance/tests/Bravo/index.test.ts
@@ -33,7 +33,7 @@ import {
   handleVotingPeriodSet,
 } from '../../src/mappings/bravo';
 import { getOrCreateDelegate } from '../../src/operations/getOrCreate';
-import { getVoteId } from '../../src/utils/ids';
+import { getVoteId } from '../../src/utilities/ids';
 import { user1 } from '../common/constants';
 import {
   createProposalCanceledEvent,
