+++ subgraphs/venus-governance/tests/Permisson/index.test.ts
@@ -9,7 +9,7 @@ import {
   handlePermissionGranted,
   handlePermissionRevoked,
 } from '../../src/mappings/accessControlManager';
-import { getPermissionId } from '../../src/utils/ids';
+import { getPermissionId } from '../../src/utilities/ids';
 import { mockContractAddress, mockFunctionSig, user1 } from '../common/constants';
 import { createPermission } from '../common/events';
